Added recycler view to the project

+++ app/src/main/java/com/artushock/moviesearcher/model/Movie.kt
@@ -3,5 +3,5 @@ package com.artushock.moviesearcher.model
 data class Movie(
         val name: String = "",
         val genre: MovieGenre = MovieGenre.DRAMA,
-        val rating: Int = 10
+        val rating: Float = 10f
 )
